api/v1/events/serializers.py

from rest_framework import serializers
from api.model.event import Event


class EventSerializer(serializers.ModelSerializer):
    class Meta:
        model = Event
        fields = "__all__"